backend/requirements.txt
asgiref==3.7.2
cffi==1.16.0
cryptography==41.0.7
Django==5.0.1
django-cleanup==8.0.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
gunicorn==21.2.0
packaging==23.2
pillow==10.2.0
psycopg2-binary==2.9.9
pycparser==2.21
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.9.0
